requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
langgraph==1.0.10
langchain==1.2.10
langchain-community==0.4.1
langchain-openai==1.1.11
tavily-python==0.7.23
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2